example(ex-gwe-bhe): Thermal loading of borehole heat exchangers (#265)
* started work on ex-gwe-bhe

* copied exampled from local file

* created py script for ex-gwe-bhe

* changed model basenames and timed functions

* working on latex docs

* added latex example description

* figure captions

* figsizes and captions

* parameters for latex table + docs

* fixed latex issues

* ruff

* lint

* added name + adjusted ex description

* added comma

---------

+++ scripts/ex-gwf-curvilinear-90.py
@@ -2165,7 +2165,7 @@ class DisvCurvilinearBuilder(DisvPropertyContainer):
             if (angle_span / angle_step) - ncol > 0.1:  # error towards larger
                 ncol += 1
         else:
-            ncol = int(round(-1 * angle_step))
+            ncol = round(-1 * angle_step)
         angle_step = angle_span / ncol
         return ncol, angle_step
